+++ editors/helix/tree-sitter-mcc/queries/textobjects.scm
@@ -1,5 +1,5 @@
 (function_definition body: (_) @function.inside) @function.around
-(extern_function) @function.around
+(function_prototype) @function.around
 
 (parameter) @parameter.inside
 (variadic_parameter) @parameter.inside